Big Update
So yeah.. For some reason a lot of my changes were still being committed
to the old repository. This should get up back on track. This fixes an
issue with the /Members route, an issue with the Domains/Edit function
and the fact that the PrestigeLogsMembershipClassesTable.php was missing
the "Table" in the filename.